requirements.txt
flask==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.5
pillow==10.0.0
werkzeug>=3.0.0
